requirements.txt
Flask==3.0.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
python-dotenv==0.19.2
Werkzeug==3.0.4
requests
